ui: style backup stronghold popup
