+++ packages/shared/components/popups/Index.svelte
@@ -7,7 +7,7 @@
     import { fade } from 'svelte/transition'
     import AddNodePopup from './AddNodePopup.svelte'
     import AirdropNetworkInfo from './AirdropNetworkInfo.svelte'
-    import Backup from './Backup.svelte'
+    import BackupStrongholdPopup from './BackupStrongholdPopup.svelte'
     import Busy from './Busy.svelte'
     import CrashReporting from './CrashReporting.svelte'
     import CreateAccountPopup from './CreateAccountPopup.svelte'
@@ -98,7 +98,7 @@
         qr: QR,
         password: Password,
         version: Version,
-        backup: Backup,
+        backupStronghold: BackupStrongholdPopup,
         deleteAccount: DeleteAccount,
         exportTransactionHistory: ExportTransactionHistory,
         hideAccount: HideAccount,
